--- Assets/Scripts/Snake/BackgroundSFX.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7c54ef00b23499f468c39e0449dfe7d1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
